+++ src/components/busTime/Departure.css
@@ -1,0 +1,13 @@
+.departure-point{
+    background: #EEEEEE;
+    border-radius: 48px ;
+    width: 93%;
+    height: 50px;
+    display: flex; /*화면중앙고정*/
+    justify-content: center;
+    align-items: center;
+    font-size: 20px;
+    line-height: 26px;   /*행간*/
+    color:black;
+  }
+
