+++ src/components/common/Footer.css
@@ -38,62 +38,3 @@ footer {
 .font-color-gray{
   color:gray;
 }
-
-.footer-bus{
-  margin:auto;
-  font-weight: bold;
-  background-color: white;
-  color:black;
-  width: 100vw;
-  height: 60px;
-  display: flex;
-  bottom: 0px;
-  position: fixed; /*이렇게하고 bottom 0px해야함*/
-  justify-content: center;
-  align-items: center;
-  font-size: 20px;
-}
-
-.footer-space{
-  margin-left: 70px;
-}
-
-.footer-to-station{
-  float:left;
-  border: white solid; 
-  border-width: 10px 0 0 0;
-  margin:auto;
-  font-weight: bold;
-  background-color: white;
-  color:black;
-  width: 50%;
-  height: 68px;
-  display: flex;
-  bottom: 0px;
-  justify-content: center;
-  align-items: center;
-  font-size: 20px;
-}
-
-.footer-to-campus{
-  border: white solid; 
-  border-width: 10px 0 0 0;
-  float:right;
-  margin:auto;
-  font-weight: bold;
-  background-color: white;
-  color:black;
-  width: 50%;
-  height: 68px;
-  display: flex;
-  bottom: 0px;
-  justify-content: center;
-  align-items: center;
-  font-size: 20px;
-}
-
-
-
-
-
-
